requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
bert-score==0.3.13
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastembed==0.5.1
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.39
langchain-google-community==2.0.7
langchain-ollama==0.2.3
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.11
llama-cloud==0.1.13
llama-cloud-services==0.6.3
llama-index-core==0.12.22
llama-parse==0.6.2
loguru==0.7.3
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.4.7
onnxruntime==1.20.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.4.0
portalocker==2.10.1
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
py_rust_stemmers==0.1.5
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
pypdf==5.3.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.1
pywin32==308
PyYAML==6.0.2
qdrant-client==1.13.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
sacrebleu==2.5.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
streamlit==1.42.2
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.23
unstructured-client==0.30.4
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
watchdog==6.0.0
webencodings==0.5.1
win32_setctime==1.2.0
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
